added mui, removed bootstrap, baseui, switched accordion/tooltips
