+++ package.json
@@ -1,8 +1,8 @@
 {
   "private": true,
   "name": "michaelkazman-portfolio",
-  "homepage": ".",
   "version": "0.1.0",
+  "homepage": ".",
   "scripts": {
     "build": "react-scripts build",
     "eject": "react-scripts eject",
@@ -11,13 +11,13 @@
   },
   "dependencies": {
     "@apollo/react-hooks": "^3.1.3",
+    "@emotion/react": "^11.4.1",
+    "@emotion/styled": "^11.3.0",
+    "@mui/material": "^5.0.2",
     "@reduxjs/toolkit": "^1.6.1",
     "apollo-boost": "^0.4.4",
-    "baseui": "^9.65.3",
-    "bootstrap": "^4.4.1",
     "chart.js": "^2.9.3",
     "fs": "0.0.1-security",
-    "gh-pages": "^2.1.1",
     "graphql": "^14.5.8",
     "node-fetch": "^2.6.0",
     "node-sass": "^4.14.1",
@@ -33,9 +33,7 @@
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0",
     "react-spring": "^8.0.27",
-    "styled-components": "^5.1.1",
-    "styletron-engine-atomic": "^1.4.6",
-    "styletron-react": "^5.2.7"
+    "styled-components": "^5.1.1"
   },
   "devDependencies": {
     "@iconify/icons-logos": "^1.0.9",
