+++ src/containers/experienceAccordion/ExperienceAccordion.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Accordion, Panel } from 'baseui/accordion';
+import { Accordion } from '@mui/material/';
 import ExperienceCard from '../../components/experienceCard/ExperienceCard';
 import './ExperienceAccordion.css';
 
@@ -7,15 +7,15 @@ const ExperienceAccordion = ({ theme, sections }) => (
   <div className="experience-accord">
     <Accordion>
       {sections.map((section) => (
-        <Panel
+        <div
           className="accord-panel"
           title={section.title}
           key={section.title}
         >
           {section.experiences.map((experience) => (
-            <ExperienceCard key={`${section.title}-experience-card`} experience={experience} theme={theme} />
+            <ExperienceCard key={`${experience.title}`} experience={experience} theme={theme} />
           ))}
-        </Panel>
+        </div>
       ))}
     </Accordion>
   </div>
